refactor : CardInfoBox 컴포넌트 분리 및 컴포넌트 파일 가독성 향상
