+++ src/InputField/CVCInputField.tsx
@@ -3,13 +3,12 @@ import BaseInputField from '../BaseInputField';
 import Input from '../Input';
 import { CVCInputValueType } from '../config/inputField';
 
-function CVCInputField({
-  inputValue,
-  setInputValue,
-}: {
+interface CVCInputFieldProps {
   inputValue: Record<CVCInputValueType, string>;
   setInputValue: Dispatch<SetStateAction<Record<CVCInputValueType, string>>>;
-}) {
+}
+
+function CVCInputField({ inputValue, setInputValue }: CVCInputFieldProps) {
   const onChange = ({ name, value }: { name: string; value: string }) => {
     if (value.length <= 3)
       setInputValue((prevValue) => ({ ...prevValue, [name]: value }));
